Moved rendering of gameroutes list to GameRoutes component
